+++ packages/blocks-core/js/libs/wordpress/post-excerpt/index.js
@@ -1,9 +1,20 @@
 // @flow
 
 /**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
  * Blockera dependencies
  */
-import { SharedBlockExtension } from '@blockera/editor';
+import {
+	SharedBlockExtension,
+	generalInnerBlockStates,
+	sharedBlockStates,
+	generalSimpleInnerBlockStates,
+} from '@blockera/editor';
+import { Icon } from '@blockera/icons';
 
 /**
  * Internal dependencies
@@ -15,7 +26,35 @@ export const PostExcerpt: BlockType = {
 	name: 'blockeraPostExcerpt',
 	targetBlock: 'core/post-excerpt',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/read-more-link': {
+			name: 'elements/read-more-link',
+			label: __('Read More Link', 'blockera'),
+			description: __('A link to read the full post.', 'blockera'),
+			icon: <Icon icon="link" library="wp" iconSize="20" />,
+			settings: {
+				force: true,
+				priority: 10,
+			},
+			availableBlockStates: {
+				...generalInnerBlockStates,
+				focus: {
+					...generalInnerBlockStates.focus,
+					force: true,
+				},
+				active: {
+					...sharedBlockStates.active,
+					force: true,
+				},
+				visited: sharedBlockStates.visited,
+			},
+		},
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
 		'elements/bold': sharedInnerBlocks['elements/bold'],
 		'elements/italic': sharedInnerBlocks['elements/italic'],
 		'elements/kbd': sharedInnerBlocks['elements/kbd'],