+++ packages/editor/js/extensions/libs/block-card/block-states/helpers.js
@@ -63,6 +63,7 @@ export function onChangeBlockStates(
 		currentBlock,
 		getStateInfo,
 		getBlockStates,
+		setCurrentBlock,
 		isMasterBlockStates,
 	} = params;
 
@@ -102,6 +103,21 @@ export function onChangeBlockStates(
 				name,
 				clientId,
 			});
+
+			const { getState, getInnerState } = select('blockera/editor');
+			const {
+				settings: { supportsInnerBlocks },
+			} = getState(stateType) ||
+				getInnerState(stateType) || {
+					settings: { supportsInnerBlocks: true },
+				};
+
+			if (
+				false === supportsInnerBlocks &&
+				'function' === typeof setCurrentBlock
+			) {
+				setCurrentBlock('master');
+			}
 		};
 
 		if (!isMasterBlockStates && state?.isSelected) {