+++ packages/blocks-core/CHANGELOG.md
@@ -52,6 +52,8 @@
 - `Verse` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Code` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Footnotes` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Preformatted` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Read More` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests